feat: Warn against malformed dict as pseudo-func arg
Enforce that if a user passes a dict as an arg to a pseudo-func like mouse(), e.g.

mouse({"x": 2000, "y": 2000})

That dict must have quotes around its keys and a colon to separate keys and values. Though we might want to auto-add quotes for the user in a future commit