feat: More modified-key mapping unicode options
This commit lets the user choose between a left, right, or
side-ambivalent modifier using the following unicode syntax:

⌘   - command
‹⌘  - left_command
⌘›  - right_command

etc. for other mods, ⌘, ⌥, ⌃, ⇧. Also added is an alias that's valid for
hyper, ☆. e.g.

☆ | 8

would map to 'command+option+control+shift + 8'

This syntax was discussed in this issue in the GokuRakuJoudo repo:
https://github.com/yqrashawn/GokuRakuJoudo/issues/205

so credit to the participants of that issue.

Addresses #2