feat: Let user define aliases for pseudo-functions
The alias must be for a valid pseudo-function, e.g. it must match the
format `funcName(args)`, where funcName is an existing pf.

This is useful if you have a bunch of clunky looking shell command
mappings next to each other or something of the sort, or if you have a
mapping you use often that you want to reference or reuse easily.