feat: Print any character with string() by falling back to AppleScript
If the string() arg contains any characters that do not have a
corresponding Karabiner key_code, then the multi-char event method
changes. Instead of sending consecutive 'to: key_code' events (which is
the fastest way), the entire string is copied to the clipboard and pasted to
the frontmost app. The prior contents of the clipboard are then restored. There
is a small delay using this AppleScript command, but in my opinon within a
tolerable range.