feat!: Let user define aliases for modifiers, and use a single line for aliases generally
BREAKING CHANGE!
Previously, the syntax for a user-defined alias was:

aliases:
  alias_name:
    - primary_key
    - [modifiers]

Now the value for an alias can use the regular syntax for a karaml map.

aliases:
  alias_name: {modifiers (optional)} {delimiter} {primary_key}

For example:

aliases:
  ⁙: ⌥ ⌃ ⇧

In the above, the ALIASES dict will be updated with `⁙` as a key and an
Alias object `Alias("left_shift", ["option", "control"])`. Moreover,
since this alias is composed entirely of modifiers, it will also be
added to the MODIFIERS_ALIAS dict as a multi-modifier. So, the following
will be available to use in any layer:

⁙ g: open(https://github.com)  # option, control, shift + g to open gh

Currently, only single character aliases are supported for modifiers.
I will work on adding support for multi-char aliases for modifiers so
that the following might be available:

rectangle_mode e: open -g "rectangle-pro://execute-action?name=left-half"