fix(feat): Distinguish delim and no-delim in aliases with all mod characters
This commit fixes a bug that converted a user defined alias with all
modifier characters, e.g. `my_alias: o c | s` into a modifier for
`left_option` + `left_control` + `left_shift`. This is a result of an
oversight from a previous feat that focused on aliases with all
symbols like `my_alias: ⌥ ⌃ ⇧`. In this case, it made sense to also
add this to the `MODIFIER_ALIASES` dict, since this alias is
composed of all aliases. However, I only considered cases like
`my_alias: o c s` as an alias for all modifiers that used the
standard single character aliases. I failed to consider that a user
might want to make an alias for `left_opt` + `left_control` + `s`,
or `m`, `c`, `o`, etc.

The solution this commit introduces is to introduce a check for one
of the valid non-whitespace delimiters, `|` or `-`. If one of those
is present in the alias definition, then the final character will
be translated as the alphabetic key it represents. So,
`my_alias: o c | s` would be added as `left_opt` + `left_control` + `s`.
If the final delimiter is only whitespace, e.g. `my_alias: o c s`, then
it is assumed the user wants to make an all-modifier alias, so that will
be translated as `left_option` + `left_control` + `left_shift`.