fix(feat): Remove restrictions on valid layer names
This commit removes the restriction that all layer names must contain
only letters. Now, any character is valid - but! there is at least one
characters, `?`, that will raise an error during Python's import of the
YAML file. This was present before this commit, and is now a known bug.
A possible solution would be to automatically wrap all the YAML keys in
single quotes when they're being loaded in by PyYAML.
The user-end workaround is to manually add single quotes around the
layer name.