fix: Prevent user from aliasing string() templates
`string()` is a template that creates a list of `to:` event. Aliases are
for singular events. The right fix is to change strings() in an alias to
the `shell_command` flavor of the string() function, which is slower but
it's a singular event. Throwing this warning in as a quick fix until then.