feat: Add descriptions parsing in layers key
This commit lets the user add an optional description in a layer,
indicated by a pipe `|` delimiter. Anything to the left of the pipe will
be parsed as the layer name, and anything to the right will be parsed as
the description. If no pipe delimiter is present, a default description
is set as it has been in prior commits, e.g. description = `/system/ layer`